+++ src/rsz/src/BaseMove.hh
@@ -14,7 +14,11 @@
 #include "db_sta/dbNetwork.hh"
 #include "db_sta/dbSta.hh"
 #include "dpl/Opendp.h"
+#include "sta/Corner.hh"
 #include "sta/FuncExpr.hh"
+#include "sta/Fuzzy.hh"
+#include "sta/Graph.hh"
+#include "sta/GraphDelayCalc.hh"
 #include "sta/MinMax.hh"
 #include "sta/Liberty.hh"
 #include "sta/PathExpanded.hh"
@@ -54,12 +58,17 @@ using sta::dbDatabase;
 using sta::dbNetwork;
 using sta::dbSta;
 using sta::DcalcAnalysisPt;
+using sta::Edge;
+using sta::fuzzyGreater;
+using sta::GraphDelayCalc;
 using sta::Instance;
+using sta::InstancePinIterator;
 using sta::InstanceSet;
 using sta::INF;
 using sta::LoadPinIndexMap;
 using sta::LibertyCell;
 using sta::LibertyPort;
+using sta::NetConnectedPinIterator;
 using sta::MinMax;
 using sta::Net;
 using sta::Network;
@@ -74,9 +83,35 @@ using sta::TimingArc;
 using sta::TimingArcSet;
 using sta::TimingRole;
 using sta::Vertex;
+using sta::VertexOutEdgeIterator;
 
 using InputSlews = std::array<Slew, RiseFall::index_count>;
 using TgtSlews = std::array<Slew, RiseFall::index_count>;
+
+struct SlackEstimatorParams
+{
+  Pin* driver_pin;
+  Pin* prev_driver_pin;
+  Pin* driver_input_pin;
+  Instance* driver;
+  const Path* driver_path;
+  const Path* prev_driver_path;
+  LibertyCell* driver_cell;
+  const float setup_slack_margin;
+  const Corner* corner;
+
+  SlackEstimatorParams(const float margin, const Corner* corner)
+      : setup_slack_margin(margin), corner(corner)
+  {
+    driver_pin = nullptr;
+    prev_driver_pin = nullptr;
+    driver_input_pin = nullptr;
+    driver = nullptr;
+    driver_path = nullptr;
+    prev_driver_path = nullptr;
+    driver_cell = nullptr;
+  }
+};
 
 class BaseMove : public sta::dbStaState {
     public:
@@ -119,7 +154,7 @@ class BaseMove : public sta::dbStaState {
         // Total accepted and pending optimizations 
         int countMoves() const;
         // Add a new pending optimization
-        void addMove(Instance* inst, bool add_count=true);
+        void addMove(Instance* inst, int count=1);
 
     protected:
        Resizer* resizer_;
@@ -130,6 +165,7 @@ class BaseMove : public sta::dbStaState {
        dbDatabase* db_ = nullptr;
        int dbu_ = 0; 
        dpl::Opendp* opendp_ = nullptr;
+       const Corner* corner_ = nullptr;
 
        // Need to track these so we don't optimize the optimzations.
        // This can result in long run-time.
@@ -187,6 +223,29 @@ class BaseMove : public sta::dbStaState {
                            const char* name,
                            Instance* parent,
                            const Point& loc);
+      bool estimatedSlackOK(const SlackEstimatorParams& params);
+      bool estimateInputSlewImpact(Instance* instance,
+                                   const DcalcAnalysisPt* dcalc_ap,
+                                   Slew old_in_slew[RiseFall::index_count],
+                                   Slew new_in_slew[RiseFall::index_count],
+                                   // delay adjustment from prev stage
+                                   float delay_adjust,
+                                   SlackEstimatorParams params,
+                                   bool accept_if_slack_improves);
+      bool hasPort(const Net* net);
+      void getBufferPins(Instance* buffer, Pin*& ip, Pin*& op);
+      int fanout(Vertex* vertex);
+
+      static constexpr int decreasing_slack_max_passes_ = 50;
+      static constexpr int rebuffer_max_fanout_ = 20;
+      static constexpr int split_load_min_fanout_ = 8;
+      static constexpr int print_interval_ = 10;
+      static constexpr int opto_small_interval_ = 100;
+      static constexpr int opto_large_interval_ = 1000;
+      static constexpr int buffer_removal_max_fanout_ = 10;
+      static constexpr float inc_fix_rate_threshold_ = 0.0001;  // default fix rate threshold = 0.01%
+      static constexpr int max_last_gasp_passes_ = 10;
+      static constexpr float rebuffer_relaxation_factor_ = 0.03;
 
 };
