+++ src/rsz/src/Rebuffer.cc
@@ -9,7 +9,8 @@
 #include <utility>
 
 #include "BufferedNet.hh"
-#include "RepairSetup.hh"
+#include "BufferMove.hh"
+//#include "RepairSetup.hh"
 #include "db_sta/dbNetwork.hh"
 #include "rsz/Resizer.hh"
 #include "sta/Corner.hh"
@@ -206,7 +207,7 @@ void RepairSetup::annotateLoadSlacks(BufferedNetPtr& bnet, Vertex* root_vertex)
 }
 
 // Find initial timing-optimized rebuffering choice
-BufferedNetPtr RepairSetup::rebufferForTiming(const BufferedNetPtr& bnet)
+BufferedNetPtr BufferMove::rebufferForTiming(const BufferedNetPtr& bnet)
 {
   using BnetType = BufferedNetType;
   using BnetSeq = BufferedNetSeq;
@@ -253,6 +254,15 @@ BufferedNetPtr RepairSetup::rebufferForTiming(const BufferedNetPtr& bnet)
           }
 
           case BufferedNetType::load: {
+            const Pin* load_pin = bnet->loadPin();
+            Vertex* vertex = graph_->pinLoadVertex(load_pin);
+            Path* req_path = sta_->vertexWorstSlackPath(vertex, resizer_->max_);
+            Path* arrival_path = nullptr;
+            if (req_path) {
+              arrival_path = req_path->prevPath();
+            }
+            bnet->setArrivalPath(arrival_path);
+            bnet->setRequiredPath(req_path);
             debugPrint(logger_,
                        RSZ,
                        "rebuffer",
@@ -301,9 +311,9 @@ BufferedNetPtr RepairSetup::rebufferForTiming(const BufferedNetPtr& bnet)
 }
 
 // Recover area on a rebuffering choice without regressing timing
-BufferedNetPtr RepairSetup::recoverArea(const BufferedNetPtr& bnet,
-                                        const Delay slack_target,
-                                        const float alpha)
+BufferedNetPtr BufferMove::recoverArea(const BufferedNetPtr& bnet,
+                                       const Delay slack_target,
+                                       const float alpha)
 {
   using BnetType = BufferedNetType;
   using BnetSeq = BufferedNetSeq;
@@ -511,7 +521,7 @@ BufferedNetPtr RepairSetup::recoverArea(const BufferedNetPtr& bnet,
   return nullptr;
 }
 
-Delay requiredDelay(const BufferedNetPtr& bnet)
+Delay BufferMove::requiredDelay(const BufferedNetPtr& bnet)
 {
   using BnetType = BufferedNetType;
   using BnetPtr = BufferedNetPtr;
@@ -539,8 +549,13 @@ Delay requiredDelay(const BufferedNetPtr& bnet)
   return worst_load_slack - bnet->slack();
 }
 
+Slack BufferMove::slackAtDriverPin(const BufferedNetPtr& bnet)
+{
+  return slackAtDriverPin(bnet, -1);
+}
+
 // Return inserted buffer count.
-int RepairSetup::rebuffer(const Pin* drvr_pin)
+int BufferMove::rebuffer(const Pin* drvr_pin)
 {
   int inserted_buffer_count = 0;
   Net* net;
@@ -669,7 +684,7 @@ Slack RepairSetup::slackAtDriverPin(const BufferedNetPtr& bnet)
 
 // Returns: driver pin gate delay; slack correction to account for changed
 // driver pin load
-std::tuple<Delay, Delay> RepairSetup::drvrPinTiming(const BufferedNetPtr& bnet)
+std::tuple<const Path*, Delay> BufferMove::drvrPinTiming(const BufferedNetPtr& bnet)
 {
   if (bnet->slackTransition() == nullptr) {
     return {0, 0};
@@ -715,7 +730,7 @@ std::tuple<Delay, Delay> RepairSetup::drvrPinTiming(const BufferedNetPtr& bnet)
   return {delay, correction};
 }
 
-Slack RepairSetup::slackAtDriverPin(const BufferedNetPtr& bnet,
+Slack BufferMove::slackAtDriverPin(const BufferedNetPtr& bnet,
                                     // Only used for debug print.
                                     const int index)
 {
@@ -735,18 +750,18 @@ Slack RepairSetup::slackAtDriverPin(const BufferedNetPtr& bnet,
 }
 
 // For testing.
-void RepairSetup::rebufferNet(const Pin* drvr_pin)
+void BufferMove::rebufferNet(const Pin* drvr_pin)
 {
   init();
   resizer_->incrementalParasiticsBegin();
-  inserted_buffer_count_ = rebuffer(drvr_pin);
+  int inserted_buffer_count_ = rebuffer(drvr_pin);
   // Leave the parasitics up to date.
   resizer_->updateParasitics();
   resizer_->incrementalParasiticsEnd();
   logger_->report("Inserted {} buffers.", inserted_buffer_count_);
 }
 
-bool RepairSetup::hasTopLevelOutputPort(Net* net)
+bool BufferMove::hasTopLevelOutputPort(Net* net)
 {
   NetConnectedPinIterator* pin_iter = network_->connectedPinIterator(net);
   while (pin_iter->hasNext()) {
@@ -760,7 +775,7 @@ bool RepairSetup::hasTopLevelOutputPort(Net* net)
   return false;
 }
 
-BufferedNetPtr RepairSetup::addWire(const BufferedNetPtr& p,
+BufferedNetPtr BufferMove::addWire(const BufferedNetPtr& p,
                                     const Point& wire_end,
                                     const int wire_layer,
                                     const int level)
@@ -793,7 +808,7 @@ BufferedNetPtr RepairSetup::addWire(const BufferedNetPtr& p,
   return z;
 }
 
-Delay RepairSetup::bufferDelay(LibertyCell* cell,
+Delay BufferMove::bufferDelay(LibertyCell* cell,
                                const RiseFallBoth* rf,
                                float load_cap)
 {
@@ -818,7 +833,7 @@ Delay RepairSetup::bufferDelay(LibertyCell* cell,
   return delay;
 }
 
-void RepairSetup::addBuffers(
+void BufferMove::addBuffers(
     BufferedNetSeq& Z1,
     const int level,
     const bool area_oriented /*=false*/,
@@ -919,12 +934,22 @@ void RepairSetup::addBuffers(
   }
 }
 
-int RepairSetup::rebufferTopDown(const BufferedNetPtr& choice,
-                                 Net* net,  // output of buffer.
-                                 const int level,
-                                 Instance* parent_in,
-                                 odb::dbITerm* mod_net_drvr,
-                                 odb::dbModNet* mod_net_in)
+float BufferMove::bufferInputCapacitance(LibertyCell* buffer_cell,
+                                          const DcalcAnalysisPt* dcalc_ap)
+{
+  LibertyPort *input, *output;
+  buffer_cell->bufferPorts(input, output);
+  const int lib_ap = dcalc_ap->libertyIndex();
+  const LibertyPort* corner_input = input->cornerPort(lib_ap);
+  return corner_input->capacitance();
+}
+
+int BufferMove::rebufferTopDown(const BufferedNetPtr& choice,
+                                Net* net,  // output of buffer.
+                                const int level,
+                                Instance* parent_in,
+                                odb::dbITerm* mod_net_drvr,
+                                odb::dbModNet* mod_net_in)
 {
   // HFix, pass in the parent
   Instance* parent = parent_in;
